requirements.txt
groq==0.13.1
langchain==0.3.13
langchain-groq==0.2.2
python-dotenv==1.0.1

